Make OVERALL_NUM_MPI_PROCS behave better (TriBITS #47)
Now, not setting OVERALL_NUM_MPI_PROCS will not automatically result in
setting the PROCESSORS property to the default number of MPI proceses when
there are only CMND test blocks.  That was very bad behavior.  Now if you
don't set OVERALL_NUM_MPI_PROCS and have all CMND test blocks, PROCESSORS is
set to 1.

Also, now the PROCESORS property is set to the max of the number of MPI
processes for all of the EXEC blocks.  Before it was ony set to the *last*
block!  That was not good behavior.

I still need to implement support for NUM_TOTAL_CORES_USED for the
TRIBITS_ADD_ADVANCED_TEST() script but that is coming next.