Prune export libraries and package lists from Config.cmake files (TriBITS #60)
This trims down the generated <Package>Config.cmake files to only contain the
direct libraries and include dirs for each SE package.

This commit is a squash of several smaller commits shown below.

1) start cleaning up the export files

2 export files: INCLUDE() dependencies in FOREACH loop

3) export files: remove redundant definition of CMAKE_CURRENT_LIST_DIR

This variable is intrinsic to CMake since 2.8.3.

4) don't use ALL dependencies in PACKAGE_LIST

We really only need the direct dependencies of the respective package.

5) typo in variable name

6) fix for Makefile exports, adapt cmake export tests

7) fix unit test

8) fix unit test 125

9) fix unit test 127