Start of refactoring to better use FIND_PACKAGE(HDF5) (TriBITS #59 and #67)
Now it only calls FIND_PACKAGE(HDF5 ...) when TPL_HDF5_INCLUDE_DIRS and
TPL_HDF5_LIBRARIES are not set.  This skips the find when there is an
override.  This will avoid some confusion I would think.

I also added logic so FIND_PACKAGE(HDF5 ...) will to also find hdf5_fortran
when HDF5_REQUIRE_FORTRAN is TRUE.  This should match existing behavior
perfectly when the user does not set any of the cache vars.  The creates a
better default find operation.

This will get refactored more as part of TriBITS #59 before it is finished.