Attempt to get the TriBITS pkg to use own TriBITS
This commit was an initial attempt to get the TriBITS package tests to use its
own implementation of TriBITS.  The idea is that we would like to be able to
have Trilinos (who has its own implementation of TriBITS snapshotted) to be
able to test an updated and different version of the TriBITS package.  The
problem is that the TriBITS snaphsot in Trilinos may not be a sufficient match
for the TriBITS in the TriBITS package.

I removed the include guards so that all TriBITS functions would get updated.

However, the version in this commit hangs after adding the test
TriBITS_TribitsWriteClientExportFiles_UnitTests.  I don't know why but I don't
have time to debug right now.