WIP: Add logic for git diffs and assert tracking branches (TriBITS #15)
I added tests to make sure I hit all the logic.

This will be used to control when to assert that all repos are on branches
with tracking branches and when we need to do git diffs looking for changed
files.

Now I need to add test cases that catch when repo is not on branch or a
tracking branch.