WIP: Grep for final PASSED/FAILED msg by default (TriBITS #15)
By default, I added a grep for the final "REQUESTED ACTIONS: PASSED" or
"REQUESTED ACTIONS: FAILED" message to make sure the checkin-test.py script
completes successfully.  I added a new optional argument
grepForFinalPassFailStr for test cases where you want the script to abort
early.

I added this because I noticed that one of the tests reported as passed
because it was supposed to fail and all of the greps matched but the script
bombed in an unexpected way and did not complete gracefully like it was
supposed to.  Adding this grep by default makes sure that the script completes
gracefully by default.