WIP: Clean up getLocalCommitsSummariesStr() (TriBITS #15)
No need to return if there are local commits or not since that is not used
anymore.

Also, I don't know the purpose of the argument appendRepoName so I took it
out.

I also trimmed down the listing of local commits for each repo.  That should
help reduce the clutter in the DID PUSH email.