WIP: Don't do git diff at all if there are no commits (TriBITS #15)
Now there will not be any git diffs at all if there are no commits for a repo.
Also, a modifiedFiles.<repoDir>.out file will not even be generated if there
are no commits for a given repo.  That just makes the script run faster when
you have a lot of repos and it avoids clutter of the CHECKIN directory.