WIP: Change push critiera to numCommitsInt(), not diff (TriBITS #15)
Now a repo is pushed to if numCommitsInt() > 0 and not if it has any diffs or
not.

This fixes a defect with the checkin-test.py script where it would not push to
a repo that has commits but no overall diff.  While this is rare, it has
happened a few times over the years.

It turns out the existing test test_extra_repo_file_3_continuous_do_all_push
had a repo that had commits but no diff.  I had to fix this test once I
chagned the behavior.  To highlight this test case, I changed the name of the
test to:

  test_extra_repo_file_3_continuous_commits_but_no_diff_do_all_push