WIP: Fix logic for enabling test dep packages (Trilinos #56)
This change I believe fixes the logic where TriBITS was enabling packages
listed in TEST_[REQUIRED|OPTIONAL]_PACKAGES even if tests and examples were
not enabled.  See trilinos/Trilinos#56.

This fails a bunch of the TriBITS tests because it now prints what type of
dependnecy ('lib' or 'test/example') the package is.  New tests targeted for
this use case need to be created and the existing failing tests need to be
fixed.