Include CMakeParseArguments module (#181)
Once we finish the full refactoring, we should, hopefully, be able to get rid
of the ParseVariableArguments.cmake module and the macro PARSE_ARGUMENTS()
altogether.

This change was needed to fix the unit tests with CMake versions prior to
3.5.0.  Interestingly, the CMakeParseArguments.cmake module got pulled in by
some of the common_tpls/utils/ helper modules by accident.  Therefore, the
rest of TriBITS was working without needing to include the module.