Disabling caching of found TPL vars (#63, ATDV-84)
For some reason, caching these vars does not set the ROL libs for SPARC.
Also, in the future I think that we may need to read in <Package>Config.cmake
everytime since it might set more vars and even create macros and functions
for downstream code to use.  Therefore, I think the next WIP commit will be to
take this out.