Add tests for DISABLED FALSE/no and fix TRIBITS_ADD_ADVANCED_TEST() (#173)
It seems logical that if the user selects

  DISABLED FALSE

or

  DISABLED NO

(or any value that equates to "false" in CMake), then it should not set the
DISABLED property and set the reason as "FASLE" or "NO".

Also, we can use this in logic for a CMake cache var
<fullTestName>_SET_DISABLED_AND_MSG that will allow a user to remove the
DISABLED test property if one configures with:

  -D <fullTestName>_SET_DISABLED_AND_MSG=FALSE

(that will be the next commit).