WIP: Add 'test_twip_2_twim_2' that 'FAILED' and make it 'PASSED'
While working on trilinos/Trilinos#3887, I noticed faulty logic in determining
global pass/fail related to 'twip' and 'twim' tests.  The existance of these
test categories should not result in a global failure.  The assumption with
'twim' is that the tests were disabled (on purpose hopefully).

To make this pass, I added the field TestsetInfo.existanceTriggersGlobalFail
that defaults to 'True' but is set to 'False' for the 'twip' and 'twim'
test-sets.