Change name of 'TestsetInfo' to 'TestsetTypeInfo' (trilinos/Trilinos#3887)
The name 'TestsetInfo' did not seem descriptive enough for me when reviewing
this.  This name makes me thing this type is describing an arbitary testset
but it is really describing a particular category of tests.  Therefore, I
think the name 'TestsetTypeInfo' makes it more clear that this type is
describing a particular type of test category.