Allow duplicate tests with same status and details (#303)
This is a very simple and safe use case to filter out duplicate tests that run
on the same day.  If all of the duplicate have the same 'status' and 'details'
fields, then any one of them should be just as good as the other really.

Hopefully this will resolve most of the problems with duplicate tests from
multiple ctest -S scripts running tests at the same time (due to backed up
queues on some systems like CTS-1).