WIP: Switch to IMPORTED INTERFACE when setting IMPORTED_LIBNAME
When setting the IMPORTED_LIBNAME property, the library target must be of type
IMPORTED INTERFACE.

This was shown out when I added a real build test case that actually loaded
these targets.

See the comments about needing to always set LD_LIBRARY_PATH to find the
shared simpletpl lib in this case.