WIP: Extend tribits_set_dep_packages() to handle TPLS
This also exposed a defect with the file
tribits/examples/MockTrilinos/TPLsList.cmake in that it did not list the TPL
'Zlib' but Zlib is listed as a TPL dependency in the file
MockTrilinos/packages/zoltan/cmake/Dependencies.cmake.  The updated code
asserted the missing definition of Zlib as a TPL and errored out.  To leave
this new behavior technically breaks backwards compatibility.  To keep
backwards compatibility, I introduced the variable:

   <Project>_ASSERT_DEFINED_DEPENDENCIES

and leaving it off by default to maintain backwards compatibility.

NOTE: If a real TriBITS project had a case like this where a listed TPL
dependency was not listed in the TPLsList.cmake file, then it would error out
when trying to access the find module thorugh ${depPkg}_FINDMOD so this is not
so harmful but it would be good to get the error message sooner.

NOTE: As part of this, a lot of tests had to be touched because the TriBITS
processing of the dependencies files no longer results in listing Zlib as a
dependency of Zoltan with MockTrilinos.  But this seems proper and correct and
the behavior otherwise does not change.