Set HAVE_<PACKAGE_NAME_UC>_<UPSTREAM_PACKAGE_NAME_UC>=OFF for missing/undefined packages (#63)
This is a better solution to the failure reported in trilinos/Trilinos#12291
and applied in the reverted commit:

  05d44294 "Only assert defined vars conditionally"
  Author: Samuel Browne <sebrown@sandia.gov>
  Date:   Thu Sep 28 13:51:05 2023 -0600 (6 days ago)

(which I had suggested at the time).

This new implementation sets
HAVE_<PACKAGE_NAME_UC>_<UPSTREAM_PACKAGE_NAME_UC>=OFF at the project level
right away when the dependencies are being read in and <upstreamPackageName>
is not defined.  This makes for more robust downstream CMake code that works
with optional upstream dependencies when those dependencies are not defined
(like when the package <upstreamPackageName> was removed from the reduced
source tree by the reduced tarball feature).  A downstream package should not
care whether an upstream dependence is disabled or is missing entirely; the
reaction should be the same and it should be able to rely on the same
variables.