Attempt to get test TriBITS_sphinx_rst_generator_UnitTests to run and pass
This test was not being run because the var HAS_PYTHON3 was not set in the
correct place so it was always disabled and printed:

  -- TriBITS_sphinx_rst_generator_UnitTests: NOT added test because EXCLUDE_IF_NOT_TRUE HAS_PYTHON3=''!

A previous commit on this branch set HAS_PYTHON3 correctly, but this test was
broken at a structural level.  The right arguments were not even being passed
into the function being unit tested (which is fixed in this commit).  It is
not clear when this test got broken because it was not being run.

The test is currently failing with the error:

  Traceback (most recent call last):
    File "/gpfs/rabartl/Trilinos.base/Trilinos/TriBITS/test/doc/sphinx_rst_generator_UnitTests.py", line 95, in test_1
      copy_file=True)
    File "/gpfs/rabartl/Trilinos.base/Trilinos/TriBITS/tribits/doc/sphinx/sphinx_rst_generator.py", line 55, in change_paths_and_get_includes
      copyfile(src=abs_path, dst=new_path, follow_symlinks=True)
    File "/usr/lib64/python3.6/shutil.py", line 120, in copyfile
      with open(src, 'rb') as fsrc:
  FileNotFoundError: [Errno 2] No such file or directory: '/gpfs/rabartl/Trilinos.base/Trilinos/TriBITS/test/examples/TribitsExampleProject/ProjectName.cmake'

I don't have time to figure this code out and fix it right now.